+++ packages/block-library/src/subhead/index.js
@@ -10,7 +10,7 @@ import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
 import save from './save';
-import transforms from './tranforms';
+import transforms from './transforms';
 
 const { name } = metadata;
