+++ packages/block-library/src/text-columns/index.js
@@ -9,7 +9,7 @@ import { __ } from '@wordpress/i18n';
 import edit from './edit';
 import metadata from './block.json';
 import save from './save';
-import transforms from './tranforms';
+import transforms from './transforms';
 
 const { name } = metadata;
