+++ packages/block-library/src/verse/index.js
@@ -11,7 +11,7 @@ import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
 import save from './save';
-import transforms from './tranforms';
+import transforms from './transforms';
 
 const { name } = metadata;
